requirements.txt
fastapi[standard]
python-jose
httpx
python-dotenv
supabase
requests
websocket-client